fix(attention): int64 seq offsets to avoid IMA on large cuTile prefill
The paged/ragged prefill kernels slice Q/K/V/O along axis 0 by
seq_start_index (query.slice(start=seq_start_index, ...)). cuda.tile computes
that view's base-pointer offset as seq_start_index * num_heads * head_dim in the
offset tensor's dtype. batch_offsets was int32, so the product overflows int32
once a tensor exceeds 2**31 elements -- an illegal memory access on large
prefills (empirically: paged total_q >= ~131072 at 128x128; ragged fails lower
because q is head_dim_qk=192, so 131072*128*192 = 3.2e9 > 2**31).

Cast batch_offsets to int64 in both entries so the base-offset math is 64-bit.
Pure addressing change (values unchanged). Verified: standard prefill suite
7 passed; large paged (100x4096, 6.7e9 out elems) and ragged (32x8192, 6.4e9
q elems) now run clean with no NaN, where they previously IMA'd.

+++ flashinfer/attention/kernels/cutile/fmha_prefill_bsr_cutile.py
@@ -1250,7 +1250,11 @@ def prefill_attention_kv_paged_cutile(
     # Flatten tensors for kernel
     actual_seq_lens_q_flat = actual_seq_lens_q.reshape(-1).contiguous()
     actual_seq_lens_kv_flat = actual_seq_lens_kv.reshape(-1).contiguous()
-    batch_offsets_flat = actual_seq_offset.reshape(-1).contiguous()
+    # int64: seq_start_index scales the axis-0 base-pointer offset of the q/o
+    # slices (seq_start * num_heads * head_dim). With int32 that product overflows
+    # once a tensor exceeds 2**31 elements (e.g. total_q>=131072 at 128x128),
+    # giving an illegal memory access on large prefills.
+    batch_offsets_flat = actual_seq_offset.reshape(-1).contiguous().to(torch.int64)
     block_tables_flat = block_tables.reshape(-1).contiguous()
     stride_block_table = block_tables.shape[1] if block_tables.dim() > 1 else 1
 
@@ -1528,7 +1532,9 @@ def prefill_attention_kv_ragged_cutile(
     # Flatten tensors for kernel
     actual_seq_lens_q_flat = actual_seq_lens_q.reshape(-1).contiguous()
     actual_seq_lens_kv_flat = actual_seq_lens_kv.reshape(-1).contiguous()
-    batch_offsets_flat = actual_seq_offset.reshape(-1).contiguous()
+    # int64: see paged entry — avoids int32 overflow of the axis-0 base-pointer
+    # offset (seq_start * num_heads * head_dim) on large ragged prefills.
+    batch_offsets_flat = actual_seq_offset.reshape(-1).contiguous().to(torch.int64)
 
     # Autotune key matching Triton's key for consistent caching across problem sizes
     autotune_key = (
